feat : 미션인증 상태 저장 위한 MissionState 객체 생성
