+++ src/main/java/com/moing/backend/domain/mission/application/service/MissionTerminationUseCase.java
@@ -4,7 +4,6 @@ import com.moing.backend.domain.member.domain.service.MemberGetService;
 import com.moing.backend.domain.mission.domain.entity.Mission;
 import com.moing.backend.domain.mission.domain.entity.constant.MissionStatus;
 import com.moing.backend.domain.mission.domain.service.MissionQueryService;
-import com.moing.backend.domain.missionStatus.domain.repository.MissionStatusRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -31,10 +30,10 @@ public class MissionTerminationUseCase {
         );
     }
 
-    // 미션 점수 반영 -> MissionStatus
+    // 미션 점수 반영 -> MissionState
 
 
-    // MissionStatus 조회 해서 미션 점수 현황조회
+    // MissionState 조회 해서 미션 점수 현황조회
 
     public void getMissionScoreStatus() {
