+++ src/main/java/com/moing/backend/domain/missionState/application/dto/MissionStatusRes.java
@@ -1,4 +1,4 @@
-package com.moing.backend.domain.missionStatus.application.dto;
+package com.moing.backend.domain.missionState.application.dto;
 
 public class MissionStatusRes {
