+++ src/main/java/com/moing/backend/domain/missionState/domain/entity/Status.java
@@ -1,0 +1,5 @@
+package com.moing.backend.domain.missionState.domain.entity;
+
+public enum Status {
+    SUCCESS,FAIL
+}
