+++ src/main/java/com/moing/backend/domain/missionState/domain/repository/MissionStateCustomRepository.java
@@ -1,0 +1,4 @@
+package com.moing.backend.domain.missionState.domain.repository;
+
+public interface MissionStateCustomRepository {
+}
