+++ src/main/java/com/moing/backend/domain/missionState/domain/repository/MissionStateRepository.java
@@ -1,0 +1,7 @@
+package com.moing.backend.domain.missionState.domain.repository;
+
+import com.moing.backend.domain.missionState.domain.entity.MissionState;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MissionStateRepository extends JpaRepository<MissionState, Long> {
+}
