--- src/main/java/com/moing/backend/domain/missionStatus/domain/entity/Status.java
@@ -1,5 +1,0 @@
-package com.moing.backend.domain.missionStatus.domain.entity;
-
-public enum Status {
-    SUCCESS,FAIL
-}
