--- src/main/java/com/moing/backend/domain/missionStatus/domain/repository/MissionStatusCustomRepository.java
@@ -1,4 +1,0 @@
-package com.moing.backend.domain.missionStatus.domain.repository;
-
-public interface MissionStatusCustomRepository {
-}
